+++ t/analysis_me_filter.t
@@ -16,11 +16,11 @@ use warnings qw{ FATAL utf8 };
 ## CPANM
 use autodie qw { :all };
 use Modern::Perl qw{ 2018 };
-use Test::Trap;
+use Readonly;
 
 ## MIPs lib/
 use lib catdir( dirname($Bin), q{lib} );
-use MIP::Constants qw{ $COLON $COMMA $SPACE set_container_constants };
+use MIP::Constants qw{ $COLON $COMMA $SPACE };
 use MIP::Test::Fixtures qw{ test_add_io_for_recipe test_log test_mip_hashes };
 
 BEGIN {
@@ -30,16 +30,16 @@ BEGIN {
 ### Check all internal dependency modules and imports
 ## Modules with import
     my %perl_module = (
-        q{MIP::Recipes::Analysis::Deepvariant} => [qw{ analysis_deepvariant }],
+        q{MIP::Recipes::Analysis::Me_filter} => [qw{ analysis_me_filter }],
         q{MIP::Test::Fixtures} => [qw{ test_add_io_for_recipe test_log test_mip_hashes }],
     );
 
     test_import( { perl_module_href => \%perl_module, } );
 }
 
-use MIP::Recipes::Analysis::Deepvariant qw{ analysis_deepvariant };
+use MIP::Recipes::Analysis::Me_filter qw{ analysis_me_filter };
 
-diag(   q{Test analysis_deepvariant from Deepvariant.pm}
+diag(   q{Test analysis_me_filter from Me_filter.pm}
       . $COMMA
       . $SPACE . q{Perl}
       . $SPACE
@@ -47,10 +47,12 @@ diag(   q{Test analysis_deepvariant from Deepvariant.pm}
       . $SPACE
       . $EXECUTABLE_NAME );
 
-test_log( { log_name => q{MIP}, no_screen => 1, } );
+Readonly my $THRESHOLD => 0.05;
+
+my $log = test_log( { log_name => q{MIP}, no_screen => 1, } );
 
 ## Given analysis parameters
-my $recipe_name    = q{deepvariant};
+my $recipe_name    = q{me_filter};
 my $slurm_mock_cmd = catfile( $Bin, qw{ data modules slurm-mock.pl } );
 
 my %active_parameter = test_mip_hashes(
@@ -59,19 +61,15 @@ my %active_parameter = test_mip_hashes(
         recipe_name   => $recipe_name,
     }
 );
-
 $active_parameter{$recipe_name}                     = 1;
 $active_parameter{recipe_core_number}{$recipe_name} = 1;
-$active_parameter{recipe_gpu_number}{$recipe_name}  = 1;
 $active_parameter{recipe_time}{$recipe_name}        = 1;
-$active_parameter{container_manager}                = q{singularity};
-my $sample_id = $active_parameter{sample_ids}[0];
-
-set_container_constants(
-    {
-        active_parameter_href => \%active_parameter,
-    }
-);
+my $case_id = $active_parameter{case_id};
+$active_parameter{sv_vcfparser_outfile_count} = 2;
+@{ $active_parameter{sv_vcfparser_select_feature_annotation_columns} } = ( 1, 2 );
+$active_parameter{sv_vcfparser_select_file} =
+  catfile( $Bin, qw{ data 643594-miptest aggregated_gene_panel_test.txt } );
+$active_parameter{me_filter_frequency_threshold} = $THRESHOLD;
 
 my %file_info = test_mip_hashes(
     {
@@ -91,24 +89,24 @@ my %parameter = test_mip_hashes(
 test_add_io_for_recipe(
     {
         file_info_href => \%file_info,
-        id             => $sample_id,
+        id             => $case_id,
         parameter_href => \%parameter,
         recipe_name    => $recipe_name,
-        step           => q{bam},
+        step           => q{vcf},
     }
 );
 
 my %sample_info;
 
-my $is_ok = analysis_deepvariant(
+my $is_ok = analysis_me_filter(
     {
         active_parameter_href => \%active_parameter,
+        case_id               => $case_id,
         file_info_href        => \%file_info,
         job_id_href           => \%job_id,
         parameter_href        => \%parameter,
         profile_base_command  => $slurm_mock_cmd,
         recipe_name           => $recipe_name,
-        sample_id             => $sample_id,
         sample_info_href      => \%sample_info,
     }
 );