+++ t/analysis_me_merge_vcfs.t
@@ -20,7 +20,7 @@ use Test::Trap;
 
 ## MIPs lib/
 use lib catdir( dirname($Bin), q{lib} );
-use MIP::Constants qw{ $COLON $COMMA $SPACE set_container_constants };
+use MIP::Constants qw{ $COLON $COMMA $SPACE };
 use MIP::Test::Fixtures qw{ test_add_io_for_recipe test_log test_mip_hashes };
 
 BEGIN {
@@ -30,16 +30,16 @@ BEGIN {
 ### Check all internal dependency modules and imports
 ## Modules with import
     my %perl_module = (
-        q{MIP::Recipes::Analysis::Deepvariant} => [qw{ analysis_deepvariant }],
+        q{MIP::Recipes::Analysis::Svdb_merge_me} => [qw{ analysis_me_merge_vcfs }],
         q{MIP::Test::Fixtures} => [qw{ test_add_io_for_recipe test_log test_mip_hashes }],
     );
 
     test_import( { perl_module_href => \%perl_module, } );
 }
 
-use MIP::Recipes::Analysis::Deepvariant qw{ analysis_deepvariant };
+use MIP::Recipes::Analysis::Svdb_merge_me qw{ analysis_me_merge_vcfs };
 
-diag(   q{Test analysis_deepvariant from Deepvariant.pm}
+diag(   q{Test analysis_me_merge_vcfs from Svdb_merge_me.pm}
       . $COMMA
       . $SPACE . q{Perl}
       . $SPACE
@@ -47,10 +47,10 @@ diag(   q{Test analysis_deepvariant from Deepvariant.pm}
       . $SPACE
       . $EXECUTABLE_NAME );
 
-test_log( { log_name => q{MIP}, no_screen => 1, } );
+test_log( { no_screen => 1, } );
 
 ## Given analysis parameters
-my $recipe_name    = q{deepvariant};
+my $recipe_name    = q{me_merge_vcfs};
 my $slurm_mock_cmd = catfile( $Bin, qw{ data modules slurm-mock.pl } );
 
 my %active_parameter = test_mip_hashes(
@@ -59,19 +59,10 @@ my %active_parameter = test_mip_hashes(
         recipe_name   => $recipe_name,
     }
 );
-
 $active_parameter{$recipe_name}                     = 1;
 $active_parameter{recipe_core_number}{$recipe_name} = 1;
-$active_parameter{recipe_gpu_number}{$recipe_name}  = 1;
 $active_parameter{recipe_time}{$recipe_name}        = 1;
-$active_parameter{container_manager}                = q{singularity};
-my $sample_id = $active_parameter{sample_ids}[0];
-
-set_container_constants(
-    {
-        active_parameter_href => \%active_parameter,
-    }
-);
+my $case_id = $active_parameter{case_id};
 
 my %file_info = test_mip_hashes(
     {
@@ -88,31 +79,39 @@ my %parameter = test_mip_hashes(
     }
 );
 
-test_add_io_for_recipe(
+SAMPLE_ID:
+foreach my $sample_id ( @{ $active_parameter{sample_ids} } ) {
+
+    test_add_io_for_recipe(
+        {
+            file_info_href => \%file_info,
+            id             => $sample_id,
+            parameter_href => \%parameter,
+            recipe_name    => $recipe_name,
+            step           => q{vcf},
+        }
+    );
+}
+
+my %sample_info = test_mip_hashes(
     {
-        file_info_href => \%file_info,
-        id             => $sample_id,
-        parameter_href => \%parameter,
-        recipe_name    => $recipe_name,
-        step           => q{bam},
+        mip_hash_name => q{qc_sample_info},
+        recipe_name   => $recipe_name,
     }
 );
 
-my %sample_info;
-
-my $is_ok = analysis_deepvariant(
+my $is_ok = analysis_me_merge_vcfs(
     {
         active_parameter_href => \%active_parameter,
+        case_id               => $case_id,
         file_info_href        => \%file_info,
         job_id_href           => \%job_id,
         parameter_href        => \%parameter,
         profile_base_command  => $slurm_mock_cmd,
         recipe_name           => $recipe_name,
-        sample_id             => $sample_id,
         sample_info_href      => \%sample_info,
     }
 );
-
 ## Then return TRUE
 ok( $is_ok, q{ Executed analysis recipe } . $recipe_name );